Deprecate all constructors and setters in CauldronRecipe, add a builder
Plans are to make this class an interface in 1.2.0 to ensure immutability. There is no reason to have this class be mutable at runtime as it's purely a data class.

The idea is that plugins depending on Alchema will upgrade before the release of 1.2.0 so that breakages aren't severe. If the builder is used, no issues should arise once this class becomes an interface.